requirements.txt
apache-airflow
pandas
sqlalchemy
psycopg2-binary
pytest
requests